fix workflows
