+++ .github/workflows/lint.yml
@@ -1,0 +1,41 @@
+# This is a basic workflow to help you get started with Actions
+
+name: Check lint
+
+# Controls when the workflow will run
+on:
+  # Triggers the workflow on push or pull request events
+  [ push, pull_request ]
+
+# A workflow run is made up of one or more jobs that can run sequentially or in parallel
+jobs:
+
+  # This workflow contains a single job called "style"
+  style:
+
+    # The type of runner that the job will run on
+    runs-on: ubuntu-latest
+
+    # A strategy creates a build matrix for your jobs
+    strategy:
+
+      # You can define a matrix of different job configurations
+      matrix:
+
+        # Each option you define in the matrix has a key and value
+        python-version: [ 3.8 ]
+
+    # Steps represent a sequence of tasks that will be executed as part of the job
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.x'
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+      - name: Pre commit
+        run: |
+          pre-commit run --all-files --color always --verbose --show-diff-on-failure
