+++ .github/workflows/mypy.yml
@@ -1,6 +1,6 @@
 # This is a basic workflow to help you get started with Actions
 
-name: Check code style
+name: Check mypy
 
 # Controls when the workflow will run
 on:
@@ -36,9 +36,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
-      - name: Run pytest
-        run: |
-          python ./scripts/run_tests.py
       - name: Run mypy
         run: |
-          mypy project/
+          mypy project/ tests/
